requirements.txt
flask
requests
requests_oauthlib
discord.py
mysql-connector-python
paramiko
spacy
python-dotenv
